+++ test/test-suite-covers-errors.sh
@@ -22,7 +22,6 @@ ${AGDA_BIN:-agda} --help=error | sed -nr 's/^([A-Z][a-z\.]+[A-Z][A-Za-z\.]+).*/\
 #
 cat > $ERRORS <<EOF
 CustomBackendError
-GenericError
 InternalError
 NonFatalErrors
 EOF